chore: renaming interface to BetterMailDTOContract
