+++ src/Resend/Email/Events/ResendEmailDeliveredEvent.php
@@ -3,7 +3,7 @@
 namespace Basement\BetterMails\Resend\Email\Events;
 
 use Basement\BetterMails\Core\Contracts\External\DeliveredEventContract;
-use Basement\BetterMails\Resend\Email\DTOs\ResendWebhookReceivedDTO;
+use Basement\BetterMails\Resend\Email\DTOs\ResendWebhookReceivedMailDTO;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
@@ -13,6 +13,6 @@ final class ResendEmailDeliveredEvent implements DeliveredEventContract
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
     public function __construct(
-        public readonly ResendWebhookReceivedDTO $dto,
+        public readonly ResendWebhookReceivedMailDTO $dto,
     ) {}
 }
