+++ src/Resend/Email/Events/ResendEmailSentEvent.php
@@ -2,7 +2,7 @@
 
 namespace Basement\BetterMails\Resend\Email\Events;
 
-use Basement\BetterMails\Resend\Email\DTOs\ResendWebhookReceivedDTO;
+use Basement\BetterMails\Resend\Email\DTOs\ResendWebhookReceivedMailDTO;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
@@ -12,6 +12,6 @@ final class ResendEmailSentEvent
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
     public function __construct(
-        public readonly ResendWebhookReceivedDTO $dto,
+        public readonly ResendWebhookReceivedMailDTO $dto,
     ) {}
 }
